I checked CRUD operations

+++ src/main/java/com/findmyclub/repositories/ClubRepository.java
@@ -3,5 +3,5 @@ package com.findmyclub.repositories;
 import com.findmyclub.model.ClubDBtable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClubRepository extends JpaRepository<ClubDBtable, Long>
+public interface ClubRepository extends JpaRepository<ClubDBtable, Integer>
 {}
